cast through uintptr_t first so we don't lose precision.  uintptr_t is only conveniently not considered a pointer type by this rev of gcc, but we're still in violation and doing it wrong.
svn:revision:38257
svn:branch:trunk
svn:account:brlcad

+++ src/libdm/dm_obj.c
@@ -3362,8 +3362,8 @@ dmo_getDrawLabelsHook_cmd(struct dm_obj *dmop,
 
     /* FIXME: the standard forbids this kind of crap.  candidate for removal. */
     sprintf(buf, "%p %p",
-	    (void *)(int)dmop->dmo_drawLabelsHook,
-	    (void *)(int)dmop->dmo_drawLabelsHookClientData);
+	    (void *)(uintptr_t)dmop->dmo_drawLabelsHook,
+	    (void *)(uintptr_t)dmop->dmo_drawLabelsHookClientData);
     Tcl_DStringInit(&ds);
     Tcl_DStringAppend(&ds, buf, -1);
     Tcl_DStringResult(interp, &ds);
